move piece utilities to its own module
